+++ components/FeaturedCars/CarCard/FeaturedWrapper/CarData.js
@@ -1,0 +1,48 @@
+import CarImage from '../../../../asset/Hero.jpg';
+
+const carCardsData = [
+  {
+    carTitle: 'Mustang Sport Car',
+    carPicture: CarImage,
+    carMake: 'Mustank',
+    carPrice: 560,
+    carLocation: 'Detroit, Michigan',
+  },
+  {
+    carTitle: 'Mustang Sport Car',
+    carPicture: CarImage,
+    carMake: 'Mustank',
+    carPrice: 560,
+    carLocation: 'Detroit, Michigan',
+  },
+  {
+    carTitle: 'Mustang Sport Car',
+    carPicture: CarImage,
+    carMake: 'Mustank',
+    carPrice: 560,
+    carLocation: 'Detroit, Michigan',
+  },
+  {
+    carTitle: 'Mustang Sport Car',
+    carPicture: CarImage,
+    carMake: 'Mustank',
+    carPrice: 560,
+    carLocation: 'Detroit, Michigan',
+  },
+  {
+    carTitle: 'Mustang Sport Car',
+    carPicture: CarImage,
+    carMake: 'Mustank',
+    carPrice: 560,
+    carLocation: 'Detroit, Michigan',
+  },
+  {
+    carTitle: 'Mustang Sport Car',
+    carPicture: CarImage,
+    carMake: 'Mustank',
+    carPrice: 560,
+    carLocation: 'Detroit, Michigan',
+  },
+];
+
+export default carCardsData;
