+++ components/FeaturedCars/CarCard/FeaturedWrapper/FeaturedWrapper.jsx
@@ -1,65 +1,21 @@
 import React from 'react';
 import Container from 'react-bootstrap/Container';
 import FeaturedCard from '../FeaturedCard';
-import CarImage from '../../../../asset/Hero.jpg';
-
-const carCardsData = [
-  {
-    carTitle: 'Mustang Sport Car',
-    carPicture: CarImage,
-    carMake: 'Mustank',
-    carPrice: 560,
-    carLocation: 'Detroit, Michigan',
-  },
-  {
-    carTitle: 'Mustang Sport Car',
-    carPicture: CarImage,
-    carMake: 'Mustank',
-    carPrice: 560,
-    carLocation: 'Detroit, Michigan',
-  },
-  {
-    carTitle: 'Mustang Sport Car',
-    carPicture: CarImage,
-    carMake: 'Mustank',
-    carPrice: 560,
-    carLocation: 'Detroit, Michigan',
-  },
-  {
-    carTitle: 'Mustang Sport Car',
-    carPicture: CarImage,
-    carMake: 'Mustank',
-    carPrice: 560,
-    carLocation: 'Detroit, Michigan',
-  },
-  {
-    carTitle: 'Mustang Sport Car',
-    carPicture: CarImage,
-    carMake: 'Mustank',
-    carPrice: 560,
-    carLocation: 'Detroit, Michigan',
-  },
-  {
-    carTitle: 'Mustang Sport Car',
-    carPicture: CarImage,
-    carMake: 'Mustank',
-    carPrice: 560,
-    carLocation: 'Detroit, Michigan',
-  },
-];
+import { Button } from 'react-bootstrap';
+import carCardsData from './CarData';
 
 const { carTitle, carPicture, carMake, carPrice, carLocation } = carCardsData;
 
 const FeaturedWrapper = () => {
   return (
-    <section className='Featured'>
+    <section className='featured--wrapper'>
       <Container>
-        <h3 className='Featured--title'>Lets Find Your Dream Car</h3>
-        <h5 className='Featured--desc'>
+        <h3 className='featured--title'>Lets Find Your Dream Car</h3>
+        <h5 className='featured--desc'>
           We recommend the very best and newest cars today and also a friendly
           price for you
         </h5>
-        <div className='Featured-Card-Wrapper'>
+        <div className='featured-card-wrapper'>
           {carCardsData.map((card) => {
             return (
               <FeaturedCard
@@ -73,10 +29,17 @@ const FeaturedWrapper = () => {
             );
           })}
         </div>
-        <button>See More</button>
+        <div>
+          <Button variant='primary' className='featured--button'>
+            See More
+          </Button>
+        </div>
       </Container>
     </section>
   );
 };
 
 export default FeaturedWrapper;
+
+// make responsive
+// calc function for rems (if columns doesn't work)
